pyproject.toml
[project]
name = "simple-auth-project"
version = "0.1.0"
description = ""
authors = [ 
    {name = "SanJeetMagar",email = "voidisstartingtocode@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular ~= 0.27.2",
    "django-extensions ~= 3.2.3",

"django-cors-headers",

"pillow (>=12.2.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
    
[dependency-groups]
dev = [
    "black (>=26.3.1,<27.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]
